Updated GPU tracking implementation. Bug fixed in cuda correlator and performance improvements

+++ src/algorithms/tracking/libs/cuda_multicorrelator.h
@@ -155,7 +155,7 @@ private:
 
     int *d_shifts_samples;
     float *d_shifts_chips;
-    float d_code_length_chips;
+    int d_code_length_chips;
 
     int threadsPerBlock;
     int blocksPerGrid;
